requirements.txt
rasa==2.4.3
rasa-sdk==2.4.1
mysql-connector-python